make PlatformViewBufferSize take a double to combat more info loss

+++ engine/src/flutter/shell/platform/android/io/flutter/embedding/engine/systemchannels/PlatformViewsChannel.java
@@ -393,12 +393,12 @@ public class PlatformViewsChannel {
   /** The platform view buffer size. */
   public static class PlatformViewBufferSize {
     /** The width of the screen buffer. */
-    public final int width;
+    public final double width;
 
     /** The height of the screen buffer. */
-    public final int height;
+    public final double height;
 
-    public PlatformViewBufferSize(int width, int height) {
+    public PlatformViewBufferSize(double width, double height) {
       this.width = width;
       this.height = height;
     }
